refactor(ui): move shared components to components/ui for clearer semantics
